update stack position

+++ lib/sites/components/songtile.dart
@@ -1,16 +1,12 @@
+import 'package:flutter/material.dart';
+
+import "../../classes/playlist.dart";
+import "../../classes/song.dart";
 import 'string_input.dart';
 import 'tagedit.dart';
-import 'package:flutter/material.dart';
-import "../../classes/song.dart";
-import "../../classes/playlist.dart";
 
-PopupMenuButton SongTile(
-    BuildContext context,
-    Song s,
-    void Function(void Function()) c,
-    MyAudioHandler Playlist,
-    bool showchild,
-    Map<int, bool> activated) {
+PopupMenuButton SongTile(BuildContext context, Song s, void Function(void Function()) c,
+    MyAudioHandler Playlist, bool showchild, Map<int, bool> activated) {
   return PopupMenuButton(
     onSelected: (result) {
       if (result == 0) {
@@ -28,11 +24,7 @@ PopupMenuButton SongTile(
                     }),
               ),
             )
-            .then((value) => {
-                  s.title = value,
-                  UpdateSongTitle(s.filename, value),
-                  c(() {})
-                });
+            .then((value) => {s.title = value, UpdateSongTitle(s.filename, value), c(() {})});
       }
       if (result == 1) {
         // Change Interpret
@@ -49,11 +41,8 @@ PopupMenuButton SongTile(
                     }),
               ),
             )
-            .then((value) => {
-                  s.interpret = value,
-                  UpdateSongInterpret(s.filename, value),
-                  c(() {})
-                });
+            .then(
+                (value) => {s.interpret = value, UpdateSongInterpret(s.filename, value), c(() {})});
       }
       if (result == 2) {
         // Change Featuring
@@ -70,11 +59,8 @@ PopupMenuButton SongTile(
                     }),
               ),
             )
-            .then((value) => {
-                  s.featuring = value,
-                  UpdateSongFeaturing(s.filename, value),
-                  c(() {})
-                });
+            .then(
+                (value) => {s.featuring = value, UpdateSongFeaturing(s.filename, value), c(() {})});
       }
       if (result == 3) {
         Navigator.of(context).push(
@@ -82,8 +68,27 @@ PopupMenuButton SongTile(
         );
       }
       if (result == 4) {
-        Playlist.RemoveSong(s);
-        DeleteSong(s);
+        showDialog(
+          context: context,
+          builder: (context) => AlertDialog(
+            title: Text("Delete " + s.title + "?"),
+            content: Text("Are you sure you want to delete this song?"),
+            actions: [
+              TextButton(
+                onPressed: () => Navigator.of(context).pop(false),
+                child: Text("No"),
+              ),
+              TextButton(
+                onPressed: () {
+                  Navigator.of(context).pop(true);
+                  Playlist.RemoveSong(s);
+                  DeleteSong(s);
+                },
+                child: Text("Yes"),
+              ),
+            ],
+          ),
+        );
       }
       if (result == 5) {
         Playlist.RemoveSong(s);
@@ -137,31 +142,25 @@ PopupMenuButton SongTile(
           value: 2,
         ),
       const PopupMenuDivider(),
-      if (activated[3] == true)
-        PopupMenuItem(child: Text('Edit Tags'), value: 3),
-      if (activated[4] == true)
-        PopupMenuItem(child: Text('Delete Song'), value: 4),
+      if (activated[3] == true) PopupMenuItem(child: Text('Edit Tags'), value: 3),
+      if (activated[4] == true) PopupMenuItem(child: Text('Delete Song'), value: 4),
       const PopupMenuDivider(),
-      if (activated[5] == true)
-        PopupMenuItem(child: Text('Play Next'), value: 5),
-      if (activated[6] == true)
-        PopupMenuItem(child: Text('Add to Playlist'), value: 6),
-      if (activated[7] == true)
-        PopupMenuItem(child: Text('Add to Stack'), value: 7),
+      if (activated[5] == true) PopupMenuItem(child: Text('Play Next'), value: 5),
+      if (activated[6] == true) PopupMenuItem(child: Text('Add to Playlist'), value: 6),
+      if (activated[7] == true) PopupMenuItem(child: Text('Add to Stack'), value: 7),
       if (activated[8] == true)
         PopupMenuItem(
-            child: Text(s.blacklisted ? 'Un Blacklist Song' : "Blacklist Song"),
-            value: 8),
+            child: Text(s.blacklisted ? 'Un Blacklist Song' : "Blacklist Song"), value: 8),
       if (activated[9] == true) PopupMenuItem(child: Text("Jump To"), value: 9),
-      if (activated[10] == true)
-        PopupMenuItem(child: Text("Move to End"), value: 10),
+      if (activated[10] == true) PopupMenuItem(child: Text("Move to End"), value: 10),
     ],
   );
 }
 
-Dismissible DismissibleSongTile(BuildContext context, Song s,
-    void Function(void Function()) c, MyAudioHandler Playlist) {
-  int i = -1; // Playlist.songs.indexOf(s); doesnt work because of the updated hash so it doesnt find the objects
+Dismissible DismissibleSongTile(
+    BuildContext context, Song s, void Function(void Function()) c, MyAudioHandler Playlist) {
+  int i =
+      -1; // Playlist.songs.indexOf(s); doesnt work because of the updated hash so it doesnt find the objects
   for (int j = 0; j < Playlist.songs.length; j++) {
     if (Playlist.songs[j].filename == s.filename) {
       i = j;
@@ -170,28 +169,30 @@ Dismissible DismissibleSongTile(BuildContext context, Song s,
   }
   if (i < 0 || i >= Playlist.songs.length) {
     return Dismissible(
-      key: Key(s.filename + s.hash + "weird"),
+      key: Key(s.filename + "weird"),
       child: ListTile(
         title: Text("ERROR: " + s.title),
         trailing: Icon(Icons.drag_handle),
-        onTap: () {
-        },
+        onTap: () {},
       ),
     );
   }
   return Dismissible(
-    key: Key(s.filename + s.hash),
+    key: Key(s.filename),
     onDismissed: (DismissDirection direction) {
       if (direction == DismissDirection.startToEnd) {
         // Dismissed to the right
         // background
         Playlist.RemoveSong(s);
-      } else {
-        // secondary background
-        s.hash += "2";
-        Playlist.Stack(s);
       }
       c(() {});
+    },
+    confirmDismiss: (dir) {
+      if (dir == DismissDirection.endToStart) {
+        Playlist.Stack(s);
+        c(() {});
+      }
+      return Future.value(true);
     },
     background: Container(
       color: Colors.red,
